movieClub/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jwt==1.3.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
tokens==0.0.3
